Update Extra - Tic Tac Toe.cpp
Hi Kevin, very happy to see you take on the challenge and even code an AI version.  One comment is that the structure of the code can be improved for readability, maybe consider separating some logic to a class and in a different file like Board class that has function for drawing and check win.  Then a player class, you can even use inheritance here, an HumanPlayer and AIPlayer.